+++ src/main/java/ulpgc/es/OptimizedMatrixMultiplication.java
@@ -2,13 +2,13 @@ package ulpgc.es;
 
 public class OptimizedMatrixMultiplication {
 
-    public static double[][] multiply(double[][] a, double[][] b) {
+    public static int[][] multiply(int[][] a, int[][] b) {
         int n = a.length;
-        double[][] result = new double[n][n];
+        int[][] result = new int[n][n];
 
         for (int i = 0; i < n; i++) {
             for (int j = 0; j < n; j++) {
-                double sum = 0;
+                int sum = 0;
                 for (int k = 0; k < n; k += 2) { // Loop unrolling
                     sum += a[i][k] * b[k][j];
                     if (k + 1 < n) {
